+++ controllers/LoginController.php
@@ -2,30 +2,14 @@
 
 namespace app\controllers;
 
+use app\controllers\AccessControllers\AnonController;
 use app\models\forms\LoginForm;
 use Yii;
-use yii\filters\AccessControl;
-use yii\helpers\Url;
-use yii\web\Controller;
 use yii\web\Response;
 use yii\widgets\ActiveForm;
 
-class LoginController extends Controller
+class LoginController extends AnonController
 {
-    public function behaviors()
-    {
-        return [
-            'access' => [
-                'class' => AccessControl::class,
-                'rules' => [
-                    [
-                        'allow' => true,
-                        'roles' => ['?']
-                    ]
-                ]
-            ]
-        ];
-    }
 
     public function actionIndex()
     {
