+++ controllers/TaskController.php
@@ -2,11 +2,10 @@
 
 namespace app\controllers;
 
+use app\controllers\AccessControllers\SecuredController;
 use app\models\forms\FilterForm;
 use app\models\Task;
 use Yii;
-use yii\db\Expression;
-use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 
 class TaskController extends SecuredController
