+++ controllers/accessControllers/SecuredController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\controllers;
+namespace app\controllers\AccessControllers;
 
 use yii\filters\AccessControl;
 use yii\web\Controller;
